Balanceamento e vida escalando com score

+++ ProjetoRoguelikeGithub/Assets/Scripts/Menu/scoreScript.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 
 public class scoreScript : MonoBehaviour
 {
-    public int pontuacao;
+    public float pontuacao;
     public TextMeshPro texto;
 
     private void Start()
